Domain 레이어 파일 위치 수정
