+++ app/src/main/java/org/sopt/official/domain/usecase/LoginUseCase.kt
@@ -24,8 +24,8 @@
  */
 package org.sopt.official.domain.usecase
 
-import domain.model.Auth
-import domain.repository.AuthRepository
+import org.sopt.officiail.domain.mypage.model.Auth
+import org.sopt.officiail.domain.mypage.repository.AuthRepository
 import javax.inject.Inject
 
 class LoginUseCase @Inject constructor(
