+++ domain/mypage/src/main/kotlin/org/sopt/officiail/domain/mypage/model/Auth.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package domain.model
+package org.sopt.officiail.domain.mypage.model
 
 data class Auth(
     val token: Token,
