+++ domain/mypage/src/main/kotlin/org/sopt/officiail/domain/mypage/model/LogOut.kt
@@ -1,4 +1,4 @@
-package domain.model
+package org.sopt.officiail.domain.mypage.model
 
 
 data class LogOut(
