+++ domain/mypage/src/main/kotlin/org/sopt/officiail/domain/mypage/model/Token.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package domain.model
+package org.sopt.officiail.domain.mypage.model
 
 data class Token(
     val accessToken: String,
